+++ Software/common/common.h
@@ -1,0 +1,15 @@
+#ifndef COMMON_H
+#define COMMON_H
+
+#include <stdio.h>
+#include <stdlib.h>
+#include <string.h>
+
+#include <fcntl.h>
+#include <unistd.h>
+#include <poll.h>
+
+
+
+
+#endif // COMMON_H
