+++ Software/src/main.c
@@ -1,41 +1,13 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <string.h>
-#include <fcntl.h>
-#include <unistd.h>
-#include <termios.h>
-#include <arpa/inet.h>
-#include <poll.h>
+#include "../common/common.h"
+#include "../inc/serialComs.h"
+
+
 
 void printUsage() {
     printf("Usage: serial_udp_bridge <serial_port> <baud_rate>\n");
 }
 
-int configureSerialPort(int serialPort, int baudRate) {
-    // Configure the serial port settings
-    struct termios serialConfig;
-    if (tcgetattr(serialPort, &serialConfig) == -1) {
-        perror("Error getting serial port attributes");
-        return -1;
-    }
-
-    cfsetispeed(&serialConfig, baudRate);
-    cfsetospeed(&serialConfig, baudRate);
-    serialConfig.c_cflag |= (CLOCAL | CREAD);
-    serialConfig.c_cflag &= ~PARENB;
-    serialConfig.c_cflag &= ~CSTOPB;
-    serialConfig.c_cflag &= ~CSIZE;
-    serialConfig.c_cflag |= CS8;
-
-    if (tcsetattr(serialPort, TCSANOW, &serialConfig) == -1) {
-        perror("Error setting serial port attributes");
-        return -1;
-    }
-
-    return 0;
-}
-
-int main(int argc, char *argv[]) {
+int main(int argc, char **argv){
     // Check if the correct number of arguments are provided
     if (argc != 3) {
         printUsage();
@@ -62,15 +34,17 @@ int main(int argc, char *argv[]) {
     pollFds[0].fd = serialPort;
     pollFds[0].events = POLLIN;
 
-    // Read from the serial port and forward to UDP
+    // Read from the serial port output to stdout 
     char buffer[256];
+    fflush(stdout);
     while (1) {
         if (poll(pollFds, 2, -1) > 0) {
             if (pollFds[0].revents & POLLIN) {
                 ssize_t bytesRead = read(serialPort, buffer, sizeof(buffer) - 1);
                 if (bytesRead > 0) {
                     buffer[bytesRead] = '\0'; // Null-terminate the string
-                    printf("%s", buffer);
+                    fputs(buffer, stdout);
+                    fflush(stdout);
                 } else if (bytesRead < 0) {
                     perror("Error reading from serial port");
                     break;
